+++ src/main/java/lk/ijse/dep11/pos/controller/ManageItemFormController.java
@@ -16,6 +16,7 @@ import javafx.scene.input.MouseEvent;
 import javafx.scene.layout.AnchorPane;
 import javafx.stage.Stage;
 import lk.ijse.dep11.pos.db.ItemDataAccess;
+import lk.ijse.dep11.pos.db.OrderDataAccess;
 import lk.ijse.dep11.pos.tm.Item;
 
 import java.io.IOException;
@@ -143,13 +144,16 @@ public class ManageItemFormController {
 
     }
     public void btnDelete_OnAction(ActionEvent actionEvent) {
-
+        Item selectedItem = tblItems.getSelectionModel().getSelectedItem();
         try {
-            Item selectedItem = tblItems.getSelectionModel().getSelectedItem();
-            // check if already associated with an order
-            ItemDataAccess.deleteItem(selectedItem.getCode());
-            tblItems.getItems().remove(selectedItem);
-            if (tblItems.getItems().isEmpty()) btnAddNew.fire();
+            if (OrderDataAccess.existsOrderByItemCode(selectedItem.getCode())){
+                new Alert(Alert.AlertType.ERROR, "Failed to delete, the item already associated with an order").show();
+            }else{
+                ItemDataAccess.deleteItem(selectedItem.getCode());
+                tblItems.getItems().remove(selectedItem);
+                if (tblItems.getItems().isEmpty()) btnAddNew.fire();
+            }
+
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
